LayoutTests/platform/win/plugins/iframe-inside-overflow-expected.txt

CONSOLE MESSAGE: PLUGIN: Plugin window rect: {left=-227, top=-189, right=73, bottom=111}
This tests that windowed plugins inside iframes inside overflow areas obey the overflow area's scroll offset.
